Refactor repository write APIs to accept typed parameters
- Introduced `MemberInfo`, `FinanceReportParams`, and `MeetupCreationParams` data classes in the `model/` package.
- Updated `UserRepository.createMember`, `TeamsRepository.addReport`, and `TeamsRepository.updateReport` to accept these strongly typed objects.
- Updated `FeedbackRepository.addReply` to accept primitive strings `(id, message, user)` instead of `JsonObject`.
- Updated `EventsRepository.createMeetup` to accept `MeetupCreationParams`.
- Moved JSON payload assembly out of `BecomeMemberActivity`, `EnterprisesViewModel`, `FeedbackDetailViewModel`, and `TeamCalendarViewModel` and directly into the respective repository implementations (`UserRepositoryImpl`, `TeamsRepositoryImpl`, `FeedbackRepositoryImpl`, and `EventsRepositoryImpl`).
- Injected `TimeProvider` into `EventsRepositoryImpl` and removed local `SystemTimeProvider` instantiations across repositories to restore deterministic date testing.
- Removed unused constructor parameters and imports from `EnterprisesViewModel`.
- Adjusted unit tests in `EventsRepositoryImplTest` and `FeedbackDetailViewModelTest` to match new method signatures and setups.
